+++ src/Organizations/OrganizationDetails/OrganizationContactInfo/OrganizationContactInfo.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { FormattedMessage } from 'react-intl';
+import { useIntl } from 'react-intl';
 
 import {
   Accordion,
@@ -13,6 +13,7 @@ import {
   hydrateAddresses,
   mixCategories,
 } from '../../../common/utils';
+import { VENDOR_CATEGORIES } from '../../../common/constants';
 
 import ContactAddresses from '../../../contacts/ViewContact/ContactAddresses';
 import ContactPersonPhones from '../../../ContactPeople/ContactPerson/ContactPersonPhones';
@@ -27,6 +28,8 @@ const filterByCatId = (catId) => ({ categories = [] }) => {
 };
 
 const OrganizationContactInfo = ({ organization, vendorCategories }) => {
+  const intl = useIntl();
+
   if (!organization) {
     return (
       <div style={{ paddingTop: '1rem' }}><Icon icon="spinner-ellipsis" width="100px" /></div>
@@ -36,7 +39,7 @@ const OrganizationContactInfo = ({ organization, vendorCategories }) => {
     ...vendorCategories,
     {
       id: UNCATEGORIZED_ID,
-      value: <FormattedMessage id="ui-organizations.contactInfo.vendorCategory.uncategorized" />,
+      value: 'Uncategorized',
     },
   ];
 
@@ -49,10 +52,10 @@ const OrganizationContactInfo = ({ organization, vendorCategories }) => {
 
     if (addresses.length || emails.length || phoneNumbers.length || urls.length) {
       acc[id] = {
-        addresses: hydrateAddresses(vendorCategories, addresses),
-        emails: mixCategories(vendorCategories, emails),
-        phoneNumbers: mixCategories(vendorCategories, phoneNumbers),
-        urls: mixCategories(vendorCategories, urls),
+        addresses: hydrateAddresses(intl, vendorCategories, addresses),
+        emails: mixCategories(intl, vendorCategories, emails),
+        phoneNumbers: mixCategories(intl, vendorCategories, phoneNumbers),
+        urls: mixCategories(intl, vendorCategories, urls),
       };
     }
 
@@ -64,7 +67,13 @@ const OrganizationContactInfo = ({ organization, vendorCategories }) => {
       <AccordionSet id="vendorCats">
         {cats.map((category, index) => data[category.id] && (
           <Accordion
-            label={category.value}
+            label={VENDOR_CATEGORIES[category.value]
+              ? intl.formatMessage({
+                id: `ui-organizations.contactInfo.vendorCategory.${VENDOR_CATEGORIES[category.value]}`,
+                defaultMessage: category.value,
+              })
+              : category.value
+            }
             id={category.id}
             closedByDefault
             key={category.id || index}