+++ src/Organizations/OrganizationDetails/OrganizationContactPeople/OrganizationContactPeople.js
@@ -1,6 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { FormattedMessage } from 'react-intl';
+import {
+  FormattedMessage,
+  useIntl,
+} from 'react-intl';
 import {
   find,
   get,
@@ -26,9 +29,12 @@ const columnMapping = {
 const alignRowProps = { alignLastColToEnd: true };
 
 const OrganizationContactPeople = ({ vendorCategories, contacts, openContact }) => {
+  const intl = useIntl();
   const resultsFormatter = {
     name: ({ firstName, lastName }) => `${lastName}, ${firstName}`,
-    categories: ({ categories = [] }) => transformCategoryIdsToLables(vendorCategories, categories) || <NoValue />,
+    categories: ({ categories = [] }) => (
+      transformCategoryIdsToLables(intl, vendorCategories, categories) || <NoValue />
+    ),
     email: ({ emails }) => get(find(emails, 'isPrimary'), 'value', '') || <NoValue />,
     phone: ({ phoneNumbers }) => get(find(phoneNumbers, 'isPrimary'), 'phoneNumber', '') || <NoValue />,
     icon: () => <Icon icon="caret-right" />,