+++ src/common/constants/index.js
@@ -4,3 +4,4 @@ export * from './interfaces';
 export * from './organization';
 export * from './routes';
 export * from './sections';
+export * from './vendorCategories';