+++ src/common/constants/vendorCategories.js
@@ -1,0 +1,7 @@
+export const VENDOR_CATEGORIES = {
+  'Customer service': 'customerService',
+  'Payments': 'payments',
+  'Returns': 'returns',
+  'Shipments': 'shipments',
+  'Uncategorized': 'uncategorized',
+};