+++ src/contacts/ViewContact/ViewContact.js
@@ -1,5 +1,8 @@
 import React from 'react';
-import { FormattedMessage } from 'react-intl';
+import {
+  FormattedMessage,
+  useIntl,
+} from 'react-intl';
 import PropTypes from 'prop-types';
 
 import {
@@ -38,8 +41,9 @@ const ViewContact = ({
   deleteContact,
 }) => {
   const [expandAll, sections, toggleSection] = useAccordionToggle();
-  const contactCategories = transformCategoryIdsToLables(categories, contact.categories);
-  const addresses = hydrateAddresses(categories, contact.addresses);
+  const intl = useIntl();
+  const contactCategories = transformCategoryIdsToLables(intl, categories, contact.categories);
+  const addresses = hydrateAddresses(intl, categories, contact.addresses);
 
   // eslint-disable-next-line react/prop-types
   const getActionMenu = ({ onToggle }) => {