+++ test/bigtest/tests/organizationDetails.e2e.js
@@ -98,9 +98,7 @@ describe('Organization details', () => {
   });
 
   it('contact people section is expanded', () => {
-    expect(
-      orgDetails.contactPeopleSection.isExpanded(orgDetails.contactPeopleSection.contentStyles),
-    ).to.be.true;
+    expect(orgDetails.contactPeopleSection.isExpanded).to.be.true;
   });
 
   describe('metadata', function () {
@@ -123,9 +121,7 @@ describe('Organization details', () => {
     });
 
     it('contact people section is closed', function () {
-      expect(
-        orgDetails.contactPeopleSection.isExpanded(orgDetails.contactPeopleSection.contentStyles),
-      ).to.be.false;
+      expect(orgDetails.contactPeopleSection.isExpanded).to.be.false;
     });
   });
 
@@ -155,9 +151,7 @@ describe('Organization details', () => {
     });
 
     it('all sections are expanded', function () {
-      expect(
-        orgDetails.contactPeopleSection.isExpanded(orgDetails.contactPeopleSection.contentStyles),
-      ).to.be.true;
+      expect(orgDetails.contactPeopleSection.isExpanded).to.be.true;
     });
   });